% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murlib.infix-package.R
\docType{package}
\name{murlib.infix-package}
\alias{murlib.infix}
\alias{murlib.infix-package}
\title{murlib.infix: Extra Infix Operators}
\description{
A re-export collection of infix operators I frequently use. These are 'extra' from base R, and
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/0000-0002-8155-2565}{ORCID})

}
\keyword{internal}
