% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{reexports}
\alias{reexports}
\alias{\%@\%}
\alias{\%@\%<-}
\alias{\%||\%}
\alias{\%|\%}
\alias{\%0\%}
\alias{\%===\%}
\alias{\%!in\%}
\alias{\%notin\%}
\title{Infix operators}
\usage{
x \%@\% name

`\%@\%`(x, name) <- value

x \%||\% y

x \%|\% y

x \%0\% y

l \%===\% r

l \%!in\% r

l \%notin\% r
}
\description{
Infix operators
}
